Move ObserveTest in its own package, make timestamped test parameterized
