+++ leshan-integration-tests/src/test/java/org/eclipse/leshan/integration/tests/observe/ObserveTest.java
@@ -17,38 +17,22 @@
  *                                                     instead of address for response
  *******************************************************************************/
 
-package org.eclipse.leshan.integration.tests;
+package org.eclipse.leshan.integration.tests.observe;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsInstanceOf.instanceOf;
 import static org.junit.Assert.*;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Set;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 
-import org.eclipse.californium.core.coap.CoAP.Type;
-import org.eclipse.californium.core.coap.OptionSet;
 import org.eclipse.californium.core.coap.Response;
-import org.eclipse.californium.core.network.CoapEndpoint;
-import org.eclipse.californium.core.network.serialization.UdpDataSerializer;
-import org.eclipse.californium.elements.AddressEndpointContext;
-import org.eclipse.californium.elements.Connector;
-import org.eclipse.californium.elements.EndpointContext;
-import org.eclipse.californium.elements.RawData;
-import org.eclipse.leshan.client.californium.LeshanClient;
-import org.eclipse.leshan.client.servers.ServerIdentity;
 import org.eclipse.leshan.core.ResponseCode;
 import org.eclipse.leshan.core.model.StaticModel;
 import org.eclipse.leshan.core.node.LwM2mObject;
 import org.eclipse.leshan.core.node.LwM2mObjectInstance;
 import org.eclipse.leshan.core.node.LwM2mPath;
 import org.eclipse.leshan.core.node.LwM2mSingleResource;
-import org.eclipse.leshan.core.node.TimestampedLwM2mNode;
 import org.eclipse.leshan.core.node.codec.LwM2mValueChecker;
 import org.eclipse.leshan.core.node.codec.json.LwM2mNodeJsonEncoder;
 import org.eclipse.leshan.core.observation.Observation;
@@ -61,8 +45,7 @@ import org.eclipse.leshan.core.response.CancelObservationResponse;
 import org.eclipse.leshan.core.response.LwM2mResponse;
 import org.eclipse.leshan.core.response.ObserveResponse;
 import org.eclipse.leshan.core.response.ReadResponse;
-import org.eclipse.leshan.server.observation.ObservationListener;
-import org.eclipse.leshan.server.registration.Registration;
+import org.eclipse.leshan.integration.tests.IntegrationTestHelper;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -168,7 +151,7 @@ public class ObserveTest {
         // verify result
         listener.waitForNotification(1000);
         assertEquals(ResponseCode.CHANGED, writeResponse.getCode());
-        assertFalse("Observation should be cancelled", listener.receivedNotify.get());
+        assertFalse("Observation should be cancelled", listener.receivedNotify().get());
     }
 
     @Test
@@ -224,7 +207,7 @@ public class ObserveTest {
         // verify result
         listener.waitForNotification(1000);
         assertEquals(ResponseCode.CHANGED, writeResponse.getCode());
-        assertFalse("Observation should be cancelled", listener.receivedNotify.get());
+        assertFalse("Observation should be cancelled", listener.receivedNotify().get());
     }
 
     @Test
@@ -328,9 +311,9 @@ public class ObserveTest {
         byte[] payload = new LwM2mNodeJsonEncoder().encode(LwM2mSingleResource.newStringResource(15, "Paris"),
                 new LwM2mPath("/3/0/15"), new StaticModel(helper.createObjectModels()), new LwM2mValueChecker());
         Response firstCoapResponse = (Response) observeResponse.getCoapResponse();
-        sendNotification(getConnector(helper.client, helper.getCurrentRegisteredServer()), payload, firstCoapResponse,
-                666); // 666 is not a supported //
-        // contentFormat.
+        // 666 is not a supported content format.
+        TestObserveUtil.sendNotification(helper.getClientConnector(helper.getCurrentRegisteredServer()),
+                helper.server.getUnsecuredAddress(), payload, firstCoapResponse, ContentFormat.fromCode(666));
         // *** Hack End *** //
 
         // verify result
@@ -339,221 +322,4 @@ public class ObserveTest {
         assertNotNull(listener.getError());
     }
 
-    @Test
-    public void can_observe_timestamped_resource() throws InterruptedException {
-        TestObservationListener listener = new TestObservationListener();
-        helper.server.getObservationService().addListener(listener);
-
-        // observe device timezone
-        ObserveResponse observeResponse = helper.server.send(helper.getCurrentRegistration(),
-                new ObserveRequest(3, 0, 15));
-        assertEquals(ResponseCode.CONTENT, observeResponse.getCode());
-        assertNotNull(observeResponse.getCoapResponse());
-        assertThat(observeResponse.getCoapResponse(), is(instanceOf(Response.class)));
-
-        // an observation response should have been sent
-        Observation observation = observeResponse.getObservation();
-        assertEquals("/3/0/15", observation.getPath().toString());
-        assertEquals(helper.getCurrentRegistration().getId(), observation.getRegistrationId());
-        Set<Observation> observations = helper.server.getObservationService()
-                .getObservations(helper.getCurrentRegistration());
-        assertTrue("We should have only on observation", observations.size() == 1);
-        assertTrue("New observation is not there", observations.contains(observation));
-
-        // *** HACK send time-stamped notification as Leshan client does not support it *** //
-        // create time-stamped nodes
-        TimestampedLwM2mNode mostRecentNode = new TimestampedLwM2mNode(System.currentTimeMillis(),
-                LwM2mSingleResource.newStringResource(15, "Paris"));
-        List<TimestampedLwM2mNode> timestampedNodes = new ArrayList<>();
-        timestampedNodes.add(mostRecentNode);
-        timestampedNodes.add(new TimestampedLwM2mNode(mostRecentNode.getTimestamp() - 2,
-                LwM2mSingleResource.newStringResource(15, "Londres")));
-        byte[] payload = new LwM2mNodeJsonEncoder().encodeTimestampedData(timestampedNodes, new LwM2mPath("/3/0/15"),
-                new StaticModel(helper.createObjectModels()), new LwM2mValueChecker());
-        Response firstCoapResponse = (Response) observeResponse.getCoapResponse();
-        sendNotification(getConnector(helper.client, helper.getCurrentRegisteredServer()), payload, firstCoapResponse,
-                ContentFormat.JSON_CODE);
-        // *** Hack End *** //
-
-        // verify result
-        listener.waitForNotification(2000);
-        assertTrue(listener.receivedNotify().get());
-        assertEquals(mostRecentNode.getNode(), listener.getResponse().getContent());
-        assertEquals(timestampedNodes, listener.getResponse().getTimestampedLwM2mNode());
-        assertNotNull(listener.getResponse().getCoapResponse());
-        assertThat(listener.getResponse().getCoapResponse(), is(instanceOf(Response.class)));
-    }
-
-    @Test
-    public void can_observe_timestamped_instance() throws InterruptedException {
-        TestObservationListener listener = new TestObservationListener();
-        helper.server.getObservationService().addListener(listener);
-
-        // observe device timezone
-        ObserveResponse observeResponse = helper.server.send(helper.getCurrentRegistration(), new ObserveRequest(3, 0));
-        assertEquals(ResponseCode.CONTENT, observeResponse.getCode());
-        assertNotNull(observeResponse.getCoapResponse());
-        assertThat(observeResponse.getCoapResponse(), is(instanceOf(Response.class)));
-
-        // an observation response should have been sent
-        Observation observation = observeResponse.getObservation();
-        assertEquals("/3/0", observation.getPath().toString());
-        assertEquals(helper.getCurrentRegistration().getId(), observation.getRegistrationId());
-        Set<Observation> observations = helper.server.getObservationService()
-                .getObservations(helper.getCurrentRegistration());
-        assertTrue("We should have only on observation", observations.size() == 1);
-        assertTrue("New observation is not there", observations.contains(observation));
-
-        // *** HACK send time-stamped notification as Leshan client does not support it *** //
-        // create time-stamped nodes
-        TimestampedLwM2mNode mostRecentNode = new TimestampedLwM2mNode(System.currentTimeMillis(),
-                new LwM2mObjectInstance(0, LwM2mSingleResource.newStringResource(15, "Paris")));
-        List<TimestampedLwM2mNode> timestampedNodes = new ArrayList<>();
-        timestampedNodes.add(mostRecentNode);
-        timestampedNodes.add(new TimestampedLwM2mNode(mostRecentNode.getTimestamp() - 2,
-                new LwM2mObjectInstance(0, LwM2mSingleResource.newStringResource(15, "Londres"))));
-        byte[] payload = new LwM2mNodeJsonEncoder().encodeTimestampedData(timestampedNodes, new LwM2mPath("/3/0"),
-                new StaticModel(helper.createObjectModels()), new LwM2mValueChecker());
-        Response firstCoapResponse = (Response) observeResponse.getCoapResponse();
-        sendNotification(getConnector(helper.client, helper.getCurrentRegisteredServer()), payload, firstCoapResponse,
-                ContentFormat.JSON_CODE);
-        // *** Hack End *** //
-
-        // verify result
-        listener.waitForNotification(2000);
-        assertTrue(listener.receivedNotify().get());
-        assertEquals(mostRecentNode.getNode(), listener.getResponse().getContent());
-        assertEquals(timestampedNodes, listener.getResponse().getTimestampedLwM2mNode());
-        assertNotNull(listener.getResponse().getCoapResponse());
-        assertThat(listener.getResponse().getCoapResponse(), is(instanceOf(Response.class)));
-    }
-
-    @Test
-    public void can_observe_timestamped_object() throws InterruptedException {
-        TestObservationListener listener = new TestObservationListener();
-        helper.server.getObservationService().addListener(listener);
-
-        // observe device timezone
-        ObserveResponse observeResponse = helper.server.send(helper.getCurrentRegistration(), new ObserveRequest(3));
-        assertEquals(ResponseCode.CONTENT, observeResponse.getCode());
-        assertNotNull(observeResponse.getCoapResponse());
-        assertThat(observeResponse.getCoapResponse(), is(instanceOf(Response.class)));
-
-        // an observation response should have been sent
-        Observation observation = observeResponse.getObservation();
-        assertEquals("/3", observation.getPath().toString());
-        assertEquals(helper.getCurrentRegistration().getId(), observation.getRegistrationId());
-        Set<Observation> observations = helper.server.getObservationService()
-                .getObservations(helper.getCurrentRegistration());
-        assertTrue("We should have only on observation", observations.size() == 1);
-        assertTrue("New observation is not there", observations.contains(observation));
-
-        // *** HACK send time-stamped notification as Leshan client does not support it *** //
-        // create time-stamped nodes
-        TimestampedLwM2mNode mostRecentNode = new TimestampedLwM2mNode(System.currentTimeMillis(),
-                new LwM2mObject(3, new LwM2mObjectInstance(0, LwM2mSingleResource.newStringResource(15, "Paris"))));
-        List<TimestampedLwM2mNode> timestampedNodes = new ArrayList<>();
-        timestampedNodes.add(mostRecentNode);
-        timestampedNodes.add(new TimestampedLwM2mNode(mostRecentNode.getTimestamp() - 2,
-                new LwM2mObject(3, new LwM2mObjectInstance(0, LwM2mSingleResource.newStringResource(15, "Londres")))));
-        byte[] payload = new LwM2mNodeJsonEncoder().encodeTimestampedData(timestampedNodes, new LwM2mPath("/3"),
-                new StaticModel(helper.createObjectModels()), new LwM2mValueChecker());
-
-        Response firstCoapResponse = (Response) observeResponse.getCoapResponse();
-        sendNotification(getConnector(helper.client, helper.getCurrentRegisteredServer()), payload, firstCoapResponse,
-                ContentFormat.JSON_CODE);
-        // *** Hack End *** //
-
-        // verify result
-        listener.waitForNotification(2000);
-        assertTrue(listener.receivedNotify().get());
-        assertEquals(mostRecentNode.getNode(), listener.getResponse().getContent());
-        assertEquals(timestampedNodes, listener.getResponse().getTimestampedLwM2mNode());
-        assertNotNull(listener.getResponse().getCoapResponse());
-        assertThat(listener.getResponse().getCoapResponse(), is(instanceOf(Response.class)));
-    }
-
-    private void sendNotification(Connector connector, byte[] payload, Response firstCoapResponse, int contentFormat) {
-
-        // create observe response
-        Response response = new Response(org.eclipse.californium.core.coap.CoAP.ResponseCode.CONTENT);
-        response.setType(Type.NON);
-        response.setPayload(payload);
-        response.setMID(firstCoapResponse.getMID() + 1);
-        response.setToken(firstCoapResponse.getToken());
-        OptionSet options = new OptionSet().setContentFormat(contentFormat)
-                .setObserve(firstCoapResponse.getOptions().getObserve() + 1);
-        response.setOptions(options);
-        EndpointContext context = new AddressEndpointContext(helper.server.getUnsecuredAddress().getAddress(),
-                helper.server.getUnsecuredAddress().getPort());
-        response.setDestinationContext(context);
-
-        // serialize response
-        UdpDataSerializer serializer = new UdpDataSerializer();
-        RawData data = serializer.serializeResponse(response);
-
-        // send it
-        connector.send(data);
-    }
-
-    private Connector getConnector(LeshanClient client, ServerIdentity server) {
-        CoapEndpoint endpoint = (CoapEndpoint) client.coap().getServer().getEndpoint(client.getAddress(server));
-        return endpoint.getConnector();
-    }
-
-    private final class TestObservationListener implements ObservationListener {
-
-        private CountDownLatch latch = new CountDownLatch(1);
-        private final AtomicBoolean receivedNotify = new AtomicBoolean();
-        private ObserveResponse response;
-        private Exception error;
-
-        @Override
-        public void onResponse(Observation observation, Registration registration, ObserveResponse response) {
-            receivedNotify.set(true);
-            this.response = response;
-            this.error = null;
-            latch.countDown();
-        }
-
-        @Override
-        public void onError(Observation observation, Registration registration, Exception error) {
-            receivedNotify.set(true);
-            this.response = null;
-            this.error = error;
-            latch.countDown();
-        }
-
-        @Override
-        public void cancelled(Observation observation) {
-            latch.countDown();
-        }
-
-        @Override
-        public void newObservation(Observation observation, Registration registration) {
-        }
-
-        public AtomicBoolean receivedNotify() {
-            return receivedNotify;
-        }
-
-        public ObserveResponse getResponse() {
-            return response;
-        }
-
-        public Exception getError() {
-            return error;
-        }
-
-        public void waitForNotification(long timeout) throws InterruptedException {
-            latch.await(timeout, TimeUnit.MILLISECONDS);
-        }
-
-        public void reset() {
-            latch = new CountDownLatch(1);
-            receivedNotify.set(false);
-            response = null;
-            error = null;
-        }
-    }
 }
